requirements.txt
alembic==1.7.1
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
docopt==0.6.2
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
passlib==1.7.4
pipreqs==0.4.10
psycopg2==2.9.1
pycparser==2.20
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.23
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
yarg==0.1.9
